Updated to revision 131 which include: Updates to PDFImage to handle JPEGs (this doesn't effect the BlackBerry version since EncodedImage takes care of this and there isn't much without writing a custom JPEG decoder that will replace EncodedImage) Handle's null images in PDFRenderer if getImage doesn't return a Bitmap. CMYK had from/toCIEXYZ implemented.
Not related to revision 131 but still in this update:
ICC_ColorSpace had fromRGB and toCIEXYZ implemented (toCIEXYZ is actually for the 131 update [see CMYKColorSpace changes])
GraphicsImpl for GraphicsInternal now has clip support, done by "changing state" so the destination graphics is actually the clip graphics, drawing the clip, then switching back to original graphics
Preparing to draw, creating mask, masking out Paint, compositing operations extracted to dedicated function (renderGraphics)
Cleanup of unused code.
Extracted the actual "draw" operation to a dedicated type (RenderEx) so that it can be used with the renderGraphics function.